move coord/feature normalizers to separate file
